cloud-init: Move `UserData` related code to userdata.go
The end goal is to make the rest of the code easier to follow by
grouping most of the low-level code generating cloud-init data files to
this userdata.go file.
